feat: more UI fixes on Gamepage and Sidebar
