+++ src/App.css
@@ -8,7 +8,7 @@
     text-align: center;
     color: var(--text-primary);
     display: grid;
-    grid-template-columns: 210px 1fr;
+    grid-template-columns: var(--sidebar-width) 1fr;
     grid-template-areas: 'sidebar content';
     font-family: 'Cabin', sans-serif;
 }
