+++ src/components/UI/Sidebar/components/SidebarLinks/index.css
@@ -1,0 +1,48 @@
+.Links {
+    font-size: 20px;
+    line-height: 23.7px;
+    font-weight: 400;
+    height: fit-content;
+    font-family: 'Rubik', sans-serif;
+    justify-content: space-between;
+    flex-direction: column;
+    display: flex;
+    grid-area: links;
+    text-align: left;
+    line-height: 36px;
+    padding-top: 24px;
+}
+
+.subItem {
+    font-size: 14px;
+    cursor: pointer;
+    padding-left: 24px;
+    line-height: 28px;
+    white-space: break-spaces;
+}
+
+.subItem:hover {
+    transition: 250ms;
+    color: var(--secondary);
+}
+
+.subItem::before {
+    content: '- ';
+}
+
+.subItem.selected {
+    color: var(--secondary)
+}
+
+.sidebar .UserSelector {
+    grid-area: user;
+}
+
+a:hover {
+    transition: 250ms;
+    color: var(--secondary);
+}
+
+.Links svg {
+    padding-right: 6px;
+}
