--- src/components/UI/Sidebar/components/SupportLinks/index.css
@@ -1,31 +1,0 @@
-.supportLinks {
-    grid-area: support;
-    justify-content: space-around;
-    align-content: center;
-    width: max-content;
-    align-self: center;
-    text-align: left;
-    justify-self: center;
-    margin-right: 16px;
-    display: flex;
-    flex-direction: column;
-    height: 50%;
-}
-
-.supportLinks>button {
-    text-align: left;
-    background: none;
-    border: solid #646464 0.1em;
-    color: white;
-    padding: 7px 12px;
-    border-radius: 2px;
-    cursor: pointer;
-    transition: 400ms;
-    text-transform: capitalize;
-    font-weight: 700;
-    margin-left: 8px;
-}
-
-.supportLinks>button:hover {
-    background: var(--background-darker);
-}
