--- src/components/UI/Sidebar/components/UserSelector/index.css
@@ -1,31 +1,0 @@
-.UserSelector {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    grid-area: user;
-    width: max-content;
-    height: fit-content;
-    border-radius: 10px;
-    margin: auto 10px;
-    transition: 250ms;
-    text-align: center;
-}
-
-.UserSelector .userName {
-    position: relative;
-    font-family: Cabin;
-    font-style: normal;
-    font-weight: 500;
-    font-size: 14px;
-    line-height: 17px;
-    color: var(--primary);
-    margin-bottom: 8px;
-    margin-left: 10%;
-    padding-right: 30px;
-    white-space: nowrap;
-}
-
-.userName>.material-icons {
-    position: relative;
-    top: 8px;
-}
