--- src/components/UI/Sidebar/components/index.tsx
@@ -1,1 +1,0 @@
-export { default as UserSelector } from './UserSelector'
