+++ src/components/UI/Sidebar/index.css
@@ -1,59 +1,16 @@
 .sidebar {
     height: 100%;
-    width: 210px;
+    width: var(--sidebar-width);
     position: fixed;
     z-index: 10;
-    background: var(--background-secondary);
+    background: var(--background);
     overflow-x: hidden;
-    padding-top: 20px;
     box-shadow: 0px 4.00313px 10.0078px #161616;
     display: grid;
-    grid-template-rows: 5fr 1fr 1fr;
-    grid-template-columns: 1fr;
-    grid-template-areas: 'links' 'support' 'user';
+    grid-template-rows: 1fr 15% 1fr;
+    grid-template-areas: 'links' 'layout' 'other';
     grid-area: sidebar;
-}
-
-.sidebar .Links {
-    font-size: 20px;
-    line-height: 23.7px;
-    font-weight: 400;
-    height: fit-content;
+    overflow: hidden;
     font-family: 'Rubik', sans-serif;
-    justify-content: space-between;
-    flex-direction: column;
-    display: flex;
-    grid-area: links;
-    text-align: left;
-    margin-left: 24px;
-    line-height: 36px;
-}
-
-.subItem {
-    font-size: 14px;
-    cursor: pointer;
-    margin-left: 14px;
-    line-height: 28px;
-    word-break: break-all;
-}
-
-.subItem:hover {
-    color: var(--secondary);
-    transition: 300ms;
-}
-
-.subItem::before {
-    content: '- ';
-}
-
-.subItem.selected {
-    color: var(--secondary)
-}
-
-.sidebar .UserSelector {
-    grid-area: user;
-}
-
-a:hover {
-    color: var(--primary);
+    padding-left: 16px;
 }
