+++ src/screens/Game/GamePage/index.css
@@ -1,19 +1,19 @@
 .gameConfigContainer {
     background: var(--background-transparent);
     box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
-    border-radius: 20px;
     margin: auto;
     display: flex;
+    grid-area: content;
+    height: 100vh;
+    width: var(--content-width);
+    justify-content: center;
 }
 
 .gameConfig {
     display: flex;
     align-self: center;
-    min-width: 780px;
-    max-width: 830px;
+    max-width: 1000px;
     margin: 44px 0 44px 44px;
-    border-right: solid 1px var(--background-transparent);
-    padding-right: 20px;
 }
 
 .gametitle {
@@ -32,7 +32,6 @@
 
 .gameInfo {
     text-align: initial;
-    padding-left: 3.5rem;
     font-size: 1rem;
     color: var(--text-primary);
     display: flex;
