+++ src/screens/Library/components/GameCard/index.css
@@ -212,7 +212,7 @@
     flex-direction: column;
     height: 100%;
     justify-content: space-around;
-    width: 120px;
+    width: 58px;
     height: 100%;
     padding: 0 6px;
 }
