+++ src/screens/Library/components/InstallModal/index.css
@@ -1,10 +1,10 @@
 .modalContainer {
     display: grid;
     place-items: center;
-    width: 100%;
+    width: var(--content-width);
     height: 100%;
     position: fixed;
-    z-index: 6;
+    z-index: 8;
 }
 
 .backdrop {
@@ -82,6 +82,10 @@
 
 .itemName {
     margin: 0 10px
+}
+
+.modalContainer>.updateIcon>svg {
+    z-index: 10;
 }
 
 .MuiIconButton-colorPrimary,
